requirements.txt
discord.py==2.7.1
python-dotenv==1.0.1
d20==1.1.2
supabase==2.28.3
